+++ src/widgets/player-history/model/type.ts
