+++ src/widgets/player-history/ui/GrowthSummaryCards.tsx
