+++ src/widgets/player-history/ui/SeasonStatsCharts.tsx
