+++ src/widgets/player-history/ui/SeasonStatsSection.tsx
