+++ src/widgets/player-history/ui/SeasonStatsTable.tsx
